packages/daifend-core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daifend-core"
version = "0.1.0"
description = "Shared models, DB, and utilities for Daifend services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "SQLAlchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "httpx>=0.27",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "ruff>=0.6"]
observability = [
  "opentelemetry-api>=1.27",
  "opentelemetry-sdk>=1.27",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-instrumentation-httpx>=0.48b0",
  "opentelemetry-exporter-otlp-proto-http>=1.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["daifend_core*"]
